Update to Youtube Vanced v13.07.55 Build 01.17.10 (minAPI21, nodpi, arm/arm64/x86/x86_64 libs)
This update includes the recent happened project renaming to "Youtube Vanced"
All references to iYTP were removed and links were updated. However to keep magisk repo functionality
repo name itself and the module id stay the same for now and will be changed at a later time.

Original iYTBP thread got closed. Support and update announcement by Team Vanced goes on here:
https://forum.xda-developers.com/android/apps-games/app-youtube-vanced-edition-t3758757

NOTE:
-----
Due to changes in the theme mod (which is switchable now!) white-stock and black-themed variant repos are
in sync now which means they are the exact same version. This will be again changed in a later update.
If u want to use the black themed variant u can use this or the other module and switch the "dark" theme via
Vanced "layout" settings. The black themed variant of the module will have the dark theme turned on by default
in the future but for now u have to switch it manually.